Move some code from the two main components
